Refactor CLI doctor code into module
